Move SortDirectionSelect from views components aggregationbuilder to widgets directory
